Refactoring: class GeoDataPointmap::EditSinglePointCommand moved to independent files.
