+++ libs/geodata/pointmap/pointmap.pro
@@ -158,6 +158,7 @@ HEADERS += gd_pointmap_global.h \
            geodatapointmapwebimporterzoomleveldialog.h \
            private/geodatapointmap_deletepointscommand.h \
            private/geodatapointmap_editpointscommand.h \
+           private/geodatapointmap_editsinglepointcommand.h \
            private/geodatapointmapcellmappert_detail.h \
            private/geodatapointmapnodemappert_detail.h \
            private/geodatapointmapt_detail.h \
@@ -195,7 +196,8 @@ SOURCES += geodatapointmap.cpp \
            geodatapointmapwebimportersettingmanager.cpp \
            geodatapointmapwebimporterzoomleveldialog.cpp \
            private/geodatapointmap_deletepointscommand.cpp \
-           private/geodatapointmap_editpointscommand.cpp
+           private/geodatapointmap_editpointscommand.cpp \
+           private/geodatapointmap_editsinglepointcommand.cpp
 RESOURCES += pointmap.qrc
 TRANSLATIONS += languages/iricGdPointmap_es_ES.ts \
                 languages/iricGdPointmap_fr_FR.ts \
